feat: 보호 라우트 처리 구현
