+++ src/app/(auth)/login/page.tsx
@@ -1,5 +1,8 @@
 import { AuthLanding } from "@/components/auth/authLanding/AuthLanding";
+import { redirectIfAuthenticated } from "@/lib/auth/guards";
 
-export default function LoginPage() {
+export default async function LoginPage() {
+  await redirectIfAuthenticated();
+
   return <AuthLanding />;
 }
