+++ src/app/profile/[username]/page.tsx
@@ -1,7 +1,11 @@
+import { requireAuthenticatedUser } from "@/lib/auth/guards";
+
 interface IProfilePageProps {
   params: { username: string };
 }
 
-export default function ProfilePage({ params }: IProfilePageProps) {
+export default async function ProfilePage({ params }: IProfilePageProps) {
+  await requireAuthenticatedUser();
+
   return <div>프로필 — @{params.username}</div>;
 }
